RDKCOM-3303: REFPLTV-1491 ctrlm-main build failure in RPi4 Hybrid 64bit
Reason for change:Build Error Test Procedure:Build successful Risks:low/none
Test Procedure: None.
Risks: Low.

(cherry picked from commit 942468e0502aaf1ef8fcb02adec7e63c1d513de6)

+++ src/voice/ipc/ctrlm_voice_ipc_request_type.h
@@ -48,7 +48,7 @@ bool ctrlm_voice_ipc_request_supported_mic_tap_stream_single(ctrlm_voice_ipc_req
 bool ctrlm_voice_ipc_request_supported_mic_tap_stream_multi(ctrlm_voice_ipc_request_config_t *config);
 bool ctrlm_voice_ipc_request_supported_mic_factory_test(ctrlm_voice_ipc_request_config_t *config);
 
-struct voice_session_request_type_handler_s *voice_session_request_type_handler_get(register const char *str, register unsigned int len);
+struct voice_session_request_type_handler_s *voice_session_request_type_handler_get(register const char *str, register size_t len);
 
 #ifdef __cplusplus
 }
